+++ node/index.ts
@@ -1,0 +1,6 @@
+import { Service } from '@vtex/api'
+
+import clients from './clients'
+import graphql from './resolvers'
+
+export default new Service({ clients, graphql })