+++ node/resolvers/index.ts
@@ -1,0 +1,9 @@
+import { saveCart } from './mutations/saveCart'
+import { getSavedCarts } from './queries/getSavedCarts'
+
+export default {
+  resolvers: {
+    Query: { getSavedCarts },
+    Mutation: { saveCart },
+  },
+}