+++ react/hooks/usePermissions.ts
@@ -1,17 +1,17 @@
 import { useMemo } from 'react'
 
-import { useOrganization } from './useOrganization'
+import { useOrganization } from '.'
 
 export function usePermissions() {
   const { organization } = useOrganization()
   const { role } = organization
 
-  const canViewMargin = useMemo(
+  const isSaleUser = useMemo(
     () =>
       !!role &&
       !['customer-admin', 'customer-approver', 'customer-buyer'].includes(role),
     [role]
   )
 
-  return { canViewMargin }
+  return { isSaleUser }
 }