+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Node and graphql setup for saved carts
+
 ## [0.0.21] - 2024-11-29
 
 ### Fixed